refactor: 불필요한 Dropdown component 제거 및 Feed Page component에 적용
