+++ frontend/src/components/DateDropdown/style.ts
@@ -2,6 +2,6 @@ import styled from "styled-components";
 
 export const Container = styled.div`
   margin-top: 10px;
-  margin-bottom: 4px;
-  width: 100%;
+  position: relative;
+  width: fit-content;
 `;
