--- frontend/src/components/Dropdown/style.ts
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  position: relative;
-  width: fit-content;
-`;
